+++ app/.server/actions/index.ts
@@ -3,4 +3,5 @@ export * from "./friendsRequestsAction";
 export * from "./groupMemberAddAction";
 export * from "./groupSettingsAction";
 export * from "./groupsCreateAction";
+export * from "./profileAction";
 export * from "./settingsAction";
